Revert "v3 Mega Update"
