+++ .gitignore
@@ -31,7 +31,6 @@ bin/
 # fabric
 
 run/
-runServer/
 
 # java
