+++ gradle.properties
@@ -3,7 +3,7 @@ org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
 # Mod Properties
-mod_version=3.0.0
+mod_version=2.1.1
 maven_group=one.devos.nautical
 archives_base_name=winterisms
