+++ src/main/java/one/devos/nautical/winterisms/config/Config.java
@@ -1,0 +1,51 @@
+package one.devos.nautical.winterisms.config;
+
+import eu.midnightdust.lib.config.MidnightConfig;
+
+public class Config extends MidnightConfig {
+    public static final String limitsGrappleCategory = "limitsGrapple";
+    public static final String flashbackCategory = "flashback";
+    public static final String limitsGrappleModId = "limits_grapple";
+    public static final String flashbackModId = "flashback";
+    public static final String onboardingCategory = "onboarding";
+    public static String painAndSufferingAndSufferingAndPain = "I understand what I am doing will cause potential instabilities with Flashback. I will not report this to Flashback support. Praise Winter. Praise Moulberry. Praise Ishland.";
+
+    @Comment(category = limitsGrappleCategory)
+    @Condition(requiredModId = limitsGrappleModId)
+    public static String limitsGrappleSwingColorWindowMissingWarning;
+
+    @Comment(category = limitsGrappleCategory)
+    @Condition(requiredModId = limitsGrappleModId)
+    public static String limitsGrappleSwingColorWindowMissingWarningSpacer;
+
+    @Entry(category = limitsGrappleCategory, isColor = true, width = 7, min = 7)
+    @Condition(requiredModId = limitsGrappleModId)
+    public static String limitsGrappleHitColor = "#3333ff";
+
+    @Entry(category = limitsGrappleCategory, isColor = true, width = 7, min = 7)
+    @Condition(requiredModId = limitsGrappleModId)
+    public static String limitsGrappleMissColor = "#99994c";
+
+    @Comment(category = flashbackCategory)
+    @Condition(requiredModId = flashbackModId)
+    public static String flashbackReplayForceAllowIncompatibleModsExtremeWarning;
+
+    @Comment(category = flashbackCategory)
+    @Condition(requiredModId = flashbackModId)
+    public static String flashbackSpacer;
+
+    @Comment(category = flashbackCategory)
+    @Condition(requiredModId = flashbackModId)
+    public static String flashbackReplayForceAllowIncompatibleModsExtremeWarning2;
+
+    @Comment(category = flashbackCategory)
+    @Condition(requiredModId = flashbackModId)
+    public static String flashbackSpacer2;
+
+    @Entry(category = flashbackCategory)
+    @Condition(requiredModId = flashbackModId)
+    public static String flashbackReplayForceAllowIncompatibleMods = "";
+
+    @Entry(category = onboardingCategory)
+    public static Boolean incompatibleModsWarningScreenViewed = false;
+}
