+++ src/main/java/one/devos/nautical/winterisms/mixin/KeyMappingAccessor.java
@@ -1,0 +1,12 @@
+package one.devos.nautical.winterisms.mixin;
+
+import com.mojang.blaze3d.platform.InputConstants;
+import net.minecraft.client.KeyMapping;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(KeyMapping.class)
+public interface KeyMappingAccessor {
+    @Accessor
+    InputConstants.Key getKey();
+}
