--- src/main/java/one/devos/nautical/winterisms/mixin/common/compat/techreborn/ElectricFurnaceBlockEntityAccessor.java
@@ -1,15 +1,0 @@
-package one.devos.nautical.winterisms.mixin.common.compat.techreborn;
-
-import net.minecraft.world.item.crafting.SmeltingRecipe;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-import techreborn.blockentity.machine.tier1.ElectricFurnaceBlockEntity;
-
-@Mixin(ElectricFurnaceBlockEntity.class)
-public interface ElectricFurnaceBlockEntityAccessor {
-    @Accessor
-    SmeltingRecipe getCurrentRecipe();
-
-    @Accessor
-    void setCurrentRecipe(SmeltingRecipe currentRecipe);
-}
