+++ src/main/java/one/devos/nautical/winterisms/mixin/compat/client/flashback/FlashbackReplayIncompatibleModsMixin.java
@@ -1,0 +1,30 @@
+package one.devos.nautical.winterisms.mixin.compat.client.flashback;
+
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import com.moulberry.flashback.Flashback;
+import com.moulberry.mixinconstraints.annotations.IfModLoaded;
+import one.devos.nautical.winterisms.config.Config;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+import java.util.List;
+import java.util.Objects;
+
+@Mixin(value = Flashback.class, priority = 1800)
+@IfModLoaded("flashback")
+public class FlashbackReplayIncompatibleModsMixin {
+    @ModifyReturnValue(
+            method = "getReplayIncompatibleMods",
+            at = @At(value = "RETURN",
+                    target = "Lcom/moulberry/flashback/Flashback;getReplayIncompatibleMods()Ljava/util/List;"
+            ),
+            remap = false
+    )
+    private static List<String> getReplayIncompatibleMods(List<String> original) {
+        if (Objects.equals(Config.flashbackReplayForceAllowIncompatibleMods, Config.painAndSufferingAndSufferingAndPain)) {
+            original.clear();
+        }
+
+        return original;
+    }
+}
