+++ src/main/java/one/devos/nautical/winterisms/mixin/compat/techreborn/ElectricFurnaceBlockEntityAccessor.java
@@ -1,0 +1,15 @@
+package one.devos.nautical.winterisms.mixin.compat.techreborn;
+
+import net.minecraft.world.item.crafting.SmeltingRecipe;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+import techreborn.blockentity.machine.tier1.ElectricFurnaceBlockEntity;
+
+@Mixin(ElectricFurnaceBlockEntity.class)
+public interface ElectricFurnaceBlockEntityAccessor {
+    @Accessor
+    SmeltingRecipe getCurrentRecipe();
+
+    @Accessor
+    void setCurrentRecipe(SmeltingRecipe currentRecipe);
+}
