--- src/main/kotlin/one/devos/nautical/winterisms/CrabInTheCode.kt
@@ -1,17 +1,0 @@
-package one.devos.nautical.winterisms
-
-/**
- * This is an extremely vital object. This object is crucial for the correct functionality of
- * Winterisms. This is extremely important, and cannot be removed under any circumstances.
- * This object single-handedly holds together the Mod. Without this object, everything
- * falls apart. This object is the reason humanity has not yet gone extinct. This object is
- * the only thing holding together the remains of the time-space continuum. Without this incredibly
- * crucial object, everything ever will fail catastrophically. In the event this object is removed,
- * expect you to go next, followed by the rest of humanity, and finally, all reality.
- * Once again, DO NOT REMOVE THIS OBJECT.
- */
-object CrabInTheCode {
-    fun crab() {
-        Winterisms.LOGGER.info("🦀")
-    }
-}
