+++ src/main/kotlin/one/devos/nautical/winterisms/client/screens/InitialIncompatibleWarningScreen.kt
@@ -1,5 +1,6 @@
 package one.devos.nautical.winterisms.client.screens
 
+import eu.midnightdust.lib.config.MidnightConfig
 import net.minecraft.ChatFormatting
 import net.minecraft.client.gui.components.MultiLineTextWidget
 import net.minecraft.client.gui.components.StringWidget
@@ -7,6 +8,7 @@ import net.minecraft.client.gui.layouts.LinearLayout
 import net.minecraft.client.gui.screens.Screen
 import net.minecraft.client.gui.screens.options.OptionsSubScreen
 import net.minecraft.network.chat.Component
+import one.devos.nautical.winterisms.Winterisms
 import one.devos.nautical.winterisms.config.Config
 import one.devos.nautical.winterisms.utils.modLoaded
 import one.devos.nautical.winterisms.utils.pairModsLoaded
@@ -98,7 +100,7 @@ class InitialIncompatibleWarningScreen(lastScreen: Screen) :
     override fun addOptions() {}
     override fun onClose() {
         Config.incompatibleModsWarningScreenViewed = true
-        Config.writeData()
+        MidnightConfig.write(Winterisms.MOD_ID)
 
         this.minecraft?.setScreen(this.lastScreen)
     }
