+++ src/main/kotlin/one/devos/nautical/winterisms/utils/ModsLoaded.kt
@@ -1,0 +1,7 @@
+package one.devos.nautical.winterisms.utils
+
+import net.fabricmc.loader.api.FabricLoader
+
+fun modLoaded(modId: String) = FabricLoader.getInstance().isModLoaded(modId)
+fun pairModsLoaded(modId1: String, modId2: String) =
+    FabricLoader.getInstance().isModLoaded(modId1) && FabricLoader.getInstance().isModLoaded(modId2)
