+++ src/main/resources/winterisms.mixins.json
@@ -7,24 +7,19 @@
     "defaultRequire": 1
   },
   "mixins": [
-    "common.CrashReportGenerativeAIDeterrentMixin",
-    "common.CrashReportMCLogsMixin",
-    "common.EndDragonFightMixin",
-    "common.compat.enderscape.MinecraftServerMixin",
-    "common.compat.journeymap.WaypointParserMixin",
-    "common.compat.techreborn.ElectricFurnaceBlockEntityAccessor",
-    "common.compat.xaeros_minimap.ClientEventsMixin"
+    "EndDragonFightMixin",
+    "compat.enderscape.MinecraftServerMixin",
+    "compat.journeymap.WaypointParserMixin",
+    "compat.techreborn.ElectricFurnaceBlockEntityAccessor",
+    "compat.xaeros_minimap.ClientEventsMixin"
   ],
   "client": [
-    "client.DisableVanillaTitle",
-    "client.KeyBindsScreenMixin",
-    "client.KeyMappingAccessor",
-    "client.TitleScreenMixin",
-    "client.WindowQOIDataMixin",
-    "client.WindowTitlebarMixin",
-    "client.compat.flashback.FlashbackReplayIncompatibleModsMixin",
-    "client.compat.flashback.FlashbackReplayIncompatibleModsTitleScreenMixin",
-    "client.compat.limits_grapple.LimitsGrappleInGameHudColorMixin",
-    "client.compat.multikeybindings.MultiKeyBindingsHandledScreenMixin"
+    "KeyBindsScreenMixin",
+    "KeyMappingAccessor",
+    "TitleScreenMixin",
+    "compat.client.flashback.FlashbackReplayIncompatibleModsMixin",
+    "compat.client.flashback.FlashbackReplayIncompatibleModsTitleScreenMixin",
+    "compat.client.limits_grapple.LimitsGrappleInGameHudColorMixin",
+    "compat.client.multikeybindings.MultiKeyBindingsHandledScreenMixin"
   ]
 }
